Summary (#536)
-Show hardware statistics.
-Show the top kernels by total duration and statistics for each kernel.
-Show individual dispatches of a chosen top kernel.
-Show aggregations of above.

+++ src/controller/src/rocprofvis_controller_node.cpp
@@ -191,7 +191,7 @@ Node::SetUInt64(rocprofvis_property_t property, uint64_t index, uint64_t value)
     {
         case kRPVControllerNodeId:
         {
-            m_id   = static_cast<uint32_t>(value);
+            m_id   = value;
             result = kRocProfVisResultSuccess;
             break;
         }
